refactor repository baseclasses
