+++ core/src/main/java/de/qabel/core/repository/ChatDropMessageRepository.kt
@@ -3,10 +3,10 @@ package de.qabel.core.repository
 import de.qabel.core.config.Contact
 import de.qabel.core.config.Identity
 import de.qabel.core.repository.entities.ChatDropMessage
-import de.qabel.core.repository.framework.BaseRepository
+import de.qabel.core.repository.framework.Repository
 import de.qabel.core.repository.framework.PagingResult
 
-interface ChatDropMessageRepository : BaseRepository<ChatDropMessage> {
+interface ChatDropMessageRepository : Repository<ChatDropMessage> {
 
     fun findByContact(contactId: Int, identityId: Int): List<ChatDropMessage>
